feat(host-internal): 迁移 CLI 标准宿主工具语义
